In this set of tasks, the focus is on pointers, arrays, and strings. The tasks involve manipulating strings, copying strings, comparing strings, reversing arrays, encoding strings into 1337 or rot13, and printing integers and buffers.
Task 0: Write a function that concatenates two strings.

Task 1: Write a function that concatenates two strings, using at most n bytes from src.

Task 2: Write a function that copies a string, using at most n bytes.

Task 3: Write a function that compares two strings.

Task 4: Write a function that reverses the content of an array of integers.

Task 5: Write a function that changes all lowercase letters of a string to uppercase.

Task 6: Write a function that capitalizes all words of a string.

Task 7: Write a function that encodes a string into 1337.

Task 8: Write a function that encodes a string using rot13.

Task 9: Write a function that prints an integer.

Task 10: Add one line of code to a given program to print a[2] = 98.

Task 11: Write a function that adds two numbers.

Task 12: Write a function that prints a buffer.